Extracts a service for code insertion logic
